scalatranslator split into several classes
