--- idealingua/idealingua-core/src/main/scala/com/github/pshirshov/izumi/idealingua/translator/toscala/extensions/CirceDerivationTranslatorExtension.scala
@@ -1,9 +1,0 @@
-package com.github.pshirshov.izumi.idealingua.translator.toscala.extensions
-
-import scala.meta._
-
-object CirceDerivationTranslatorExtension extends CirceTranslatorExtensionBase {
-  override protected val classDeriverImports: List[Import] = List(
-    q""" import _root_.io.circe.derivation.{deriveDecoder, deriveEncoder} """
-  )
-}
