+++ idealingua/idealingua-core/src/test/scala/com/github/pshirshov/izumi/idealingua/CompilerTest.scala
@@ -1,6 +1,8 @@
 package com.github.pshirshov.izumi.idealingua
 
 import com.github.pshirshov.izumi.fundamentals.platform.files.IzFiles
+import com.github.pshirshov.izumi.idealingua.translator.toscala.ScalaTranslator
+import com.github.pshirshov.izumi.idealingua.translator.toscala.extensions.{CirceDerivationTranslatorExtension, CirceGenericTranslatorExtension}
 import org.scalatest.WordSpec
 
 
@@ -13,20 +15,33 @@ class CompilerTest extends WordSpec {
       assume(IzFiles.haveExecutables("scalac"), "scalac not available")
       assert(compilesScala(getClass.getSimpleName, loadDefs()))
     }
+
     "be able to compile into typescript" in {
       assume(IzFiles.haveExecutables("tsc"), "tsc not available")
       assume(IzFiles.haveExecutables("npm"), "tsc not available")
       assert(compilesTypeScript(getClass.getSimpleName, loadDefs(), scoped = false))
       assert(compilesTypeScript(getClass.getSimpleName, loadDefs(), scoped = true))
     }
+
     "be able to compile into golang" in {
       assume(IzFiles.haveExecutables("go"), "go not available")
       assert(compilesGolang(getClass.getSimpleName, loadDefs(), scoped = false))
       assert(compilesGolang(getClass.getSimpleName, loadDefs(), scoped = true))
     }
+
     "be able to compile into csharp" in {
       assume(IzFiles.haveExecutables("csc", "nunit-console"), "csc not available")
       assert(compilesCSharp(getClass.getSimpleName, loadDefs()))
+    }
+
+    "be able to compile into scala with circe-derivation" in {
+      assume(IzFiles.haveExecutables("scalac"), "scalac not available")
+      assert(compilesScala(getClass.getSimpleName, loadDefs(), ScalaTranslator.defaultExtensions ++ Seq(CirceDerivationTranslatorExtension)))
+    }
+
+    "be able to compile into scala with circe-generic" in {
+      assume(IzFiles.haveExecutables("scalac"), "scalac not available")
+      assert(compilesScala(getClass.getSimpleName, loadDefs(), ScalaTranslator.defaultExtensions ++ Seq(CirceGenericTranslatorExtension)))
     }
   }
 }
