--- idealingua/idealingua-core/src/test/scala/com/github/pshirshov/izumi/idealingua/CompilerWithCirceTest.scala
@@ -1,23 +1,0 @@
-package com.github.pshirshov.izumi.idealingua
-
-import com.github.pshirshov.izumi.fundamentals.platform.files.IzFiles
-import com.github.pshirshov.izumi.idealingua.translator.toscala.extensions.{CirceDerivationTranslatorExtension, CirceGenericTranslatorExtension}
-import com.github.pshirshov.izumi.idealingua.translator.toscala.ScalaTranslator
-import org.scalatest.WordSpec
-
-class CompilerWithCirceTest extends WordSpec {
-
-  import IDLTestTools._
-
-  "IDL compiler" should {
-    "be able to compile into scala with circe-derivation" in {
-      assume(IzFiles.haveExecutables("scalac"), "scalac not available")
-      assert(compilesScala(getClass.getSimpleName, loadDefs(), ScalaTranslator.defaultExtensions ++ Seq(CirceDerivationTranslatorExtension)))
-    }
-
-    "be able to compile into scala with circe-generic" in {
-      assume(IzFiles.haveExecutables("scalac"), "scalac not available")
-      assert(compilesScala(getClass.getSimpleName, loadDefs(), ScalaTranslator.defaultExtensions ++ Seq(CirceGenericTranslatorExtension)))
-    }
-  }
-}
